Fix bug in the imports paths after move the folders
