+++ src/core/main/use-initialice-top.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 //own
-import { getTopRated } from "store/collections/top/get-top-rated";
+import { getTopRated } from "store/top/get-top-rated";
 export default function useInitialiceTopRated() {
   const dispatch = useDispatch();
