+++ src/core/main/use-initialice-tv.js
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 
 //own
-import { getTvPopular } from "store/collections/tv/get-tv-popular";
+import { getTvPopular } from "store/tv/get-tv-popular";
 export default function useInitialiceTv() {
   const dispatch = useDispatch();
