+++ src/store/index.js
@@ -1,10 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import moviesSlice from "store/movies/movies-slice";
-import tvSlice from "store/collections/tv/tv-slice";
+import tvSlice from "store/tv/tv-slice";
 import genresSlice from "store/genres/genres-slice";
 import detailSlice from "store/details/detail-slice";
 import searchSlice from "store/search/search-slice";
-import topRatedSlice from "store/collections/top/top-rated-slice";
+import topRatedSlice from "store/top/top-rated-slice";
 import advancesSlice from "store/advances/advances-slice";
 
 const store = configureStore({
